huffman coding: elementary commit.
This is the first commit message for huffman coder. Currently there is
facility only to generate huffman codes for a given text file. Huffman
codes are generated for first 128 ASCII characters. You need to give
filename path and then have to compile the code to successfully generate
huffman table on console.